CRUD for Squads and Places + UserDAO

+++ Server/src/main/java/com/momot/trakball/repository/SquadRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SquadRepository extends CrudRepository<Squad, Integer> {
+public interface SquadRepository extends CrudRepository<Squad, Long> {
